+++ NEWS.md
@@ -1,3 +1,8 @@
+# LAIr 0.2.0
+* Fixed an error on non-handling NAs.
+* Fixed an error on filtering by ID.
+* Added `biome` as argument for filtering.
+
 # LAIr 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.